requirements.txt
langchain[google-genai]
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
starlette
duckduckgo-search
ddgs
bs4